[UPUP-986] разбиение по отдельным файлам
