+++ Sources/NetworkService/BaseNetworkService.swift
@@ -84,52 +84,3 @@ open class BaseNetworkService<Target: MobileApiTargetType> {
         }
     }
 }
-
-private extension BaseNetworkService {
-    actor TokenRefresher {
-        private let tokenRefreshProvider: TokenRefreshProvider
-        private var refreshTokenTask: _Concurrency.Task<Void, Error>?
-
-        init(tokenRefreshProvider: TokenRefreshProvider) {
-            self.tokenRefreshProvider = tokenRefreshProvider
-        }
-
-        func refreshToken() async throws {
-            Log.refreshTokenFlow.debug(logEntry: .text("NetworkService. RefreshToken method called"))
-
-            if let task = refreshTokenTask {
-                return try await task.value
-            }
-
-            refreshTokenTask = _Concurrency.Task { [weak self] in
-                guard let self else { throw CancellationError() }
-
-                Log.refreshTokenFlow.debug(logEntry: .text("NetworkService. RefreshToken request started"))
-
-                do {
-                    _ = try await tokenRefreshProvider.refreshToken()
-                    Log.refreshTokenFlow.debug(logEntry: .text("NetworkService. RefreshToken updated"))
-                } catch {
-                    Log.refreshTokenFlow.debug(logEntry: .text("NetworkService. RefreshToken failed: \(error)"))
-                    throw error
-                }
-            }
-
-            try await refreshTokenTask?.value
-        }
-    }
-
-    actor OnceExecutor {
-        private var hasRun = false
-        private var onTokenRefreshFailed: (() -> Void)?
-
-        func executeTokenRefreshFailed() async {
-            guard hasRun == false else {
-                return
-            }
-            hasRun = true
-            onTokenRefreshFailed?()
-            Log.refreshTokenFlow.debug(logEntry: .text("NetworkService. Send onTokenRefreshFailed"))
-        }
-    }
-}
